refactor: split the event-channel payloads out of the channel registry
`channels.ts` was 474 lines doing two jobs. The registry half answers "which
channels exist, and what state does each one's last-value replay cache"; the
other half is the request BODIES the event channels carry — `tab.open` asks a
TUI to open a pane, `ui.prompt` asks a human for a line of text,
`notice.event` asks every attached UI for a toast. A consumer dedupes each of
those on its `at` instead of rendering it as current state, which is why they
read badly inside the registry's state-channel narrative.

Pure type movement into `channels-events.ts`, re-exported from `channels.ts`,
so `protocol.ts` remains the single public import path and no consumer changes.

  before: wc -l packages/kobe-daemon/src/daemon/channels.ts -> 474
  after:  387 + 122 (channels-events.ts)
