+++ packages/kobe-daemon/src/daemon/channels.ts
@@ -8,6 +8,15 @@
 
 import { DAEMON_CHANNELS } from "@sma1lboy/rove-plugin-sdk/contract"
 import type {
+  EngineLifecyclePayload,
+  NoticeEventPayload,
+  SessionDeliverPayload,
+  TabClosePayload,
+  TabOpenPayload,
+  TabRenamePayload,
+  UiPromptPayload,
+} from "./channels-events.ts"
+import type {
   AttentionInboxItem,
   EngineActivityDetail,
   EngineContextUsage,
@@ -15,6 +24,17 @@ import type {
   TaskActivityState,
   UpdateInfo,
 } from "./contracts.ts"
+export type {
+  EngineLifecyclePayload,
+  NoticeEventPayload,
+  PaneClosePayload,
+  SessionDeliverPayload,
+  TabClosePayload,
+  TabOpenPayload,
+  TabRenamePayload,
+  TerminalTabClosePayload,
+  UiPromptPayload,
+} from "./channels-events.ts"
 import type { RepoIssues } from "./issues-store.ts"
 import type { SerializedTask } from "./protocol.ts"
 
@@ -115,7 +135,11 @@ export interface ChannelPayloads {
    * file.
    */
   "ui-prefs": {
-    theme: string
+    /** Selected theme NAME, or `null` when `state.json` names none. The daemon
+     *  has no theme registry, so it cannot supply a default — a literal here
+     *  would be a second, silently-drifting copy of the TUI's. `null` means
+     *  "no opinion": `applyUiPrefs` leaves whatever theme the pane already has. */
+    theme: string | null
     transparentBackground: boolean
     focusAccent: string | null
     /** UI language id (`state.json`'s `locale`). Opaque to the daemon — the TUI validates it. */
@@ -309,113 +333,6 @@ export interface ChannelPayloads {
   // `client.onChannel(name, …)` in a consumer — that's the whole recipe:
   // "cost": { taskId: string; usd: number; tokens: number }
   // "pr-status": { taskId: string; state: "open" | "merged" | "closed" | "none" }
-}
-
-/** The `notice.event` channel payload — one toast for every attached UI. */
-export interface NoticeEventPayload {
-  readonly title: string
-  /** Optional second line under the title — context, not a second message. */
-  readonly body?: string
-  /**
-   * Free-form kind tag. The TUI styles the known severities
-   * ("done" / "needs_input" / "error" — its NotificationKind vocabulary)
-   * and renders anything else neutrally, so agents may invent their own.
-   */
-  readonly kind: string
-  /** Optional task the notice concerns (drives the sidebar unread mark). */
-  readonly taskId?: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-  /** Free-form origin tag (e.g. "api", an agent name). */
-  readonly source?: string
-}
-
-/** The `session.deliver` channel payload — one "paste this into task X". */
-export interface SessionDeliverPayload {
-  readonly taskId: string
-  readonly text: string
-  /** Exact terminal tab to deliver into (`dispatch --tab`); absent = the
-   *  canonical engine tab. */
-  readonly tabId?: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-  readonly source: "note" | "dispatcher"
-}
-
-/** The `engine.lifecycle` channel payload — one low-frequency agent-lifecycle signal. */
-export interface EngineLifecyclePayload {
-  readonly taskId: string
-  readonly kind: "pre-compact" | "post-compact" | "subagent-start" | "subagent-stop"
-  readonly tabId?: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-}
-
-/** The `tab.open` channel payload — one "open a terminal pane running argv". */
-export interface TabOpenPayload {
-  readonly taskId: string
-  /** Argv the pane's PTY spawns verbatim (no shell wrap on this side). */
-  readonly argv: readonly string[]
-  readonly title: string
-  /** Host tab for the split (`pane-open --tab`); absent = the focused tab. */
-  readonly tabId?: string
-  /** `split` (default) joins the focused Terminal Tab's split group; `tab` opens a separate tab. */
-  readonly placement?: "split" | "tab"
-  /** Split orientation: `right` (default) lays the new pane beside the
-   *  active leaf, `down` stacks it below. Ignored for `placement: "tab"`. */
-  readonly direction?: "right" | "down"
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-}
-
-/** The `tab.close` channel's pane-close variant. */
-export interface PaneClosePayload {
-  readonly taskId: string
-  /** Pane label to close — matches the `title` split leaves / command tabs
-   *  were opened with (`tab.open`); engine leaves are never closed. */
-  readonly title: string
-  /** Scope the title match to one tab (`pane-close --tab`); absent = all
-   *  tabs of the task. */
-  readonly tabId?: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-}
-
-/** The `tab.close` channel's exact Terminal Tab close variant. */
-export interface TerminalTabClosePayload {
-  readonly kind: "terminal-tab"
-  readonly taskId: string
-  readonly tabId: string
-  /** Correlates the TUI's close result with the waiting CLI request. */
-  readonly requestId: string
-  readonly at: number
-}
-
-/** Pane closes retain their existing wire shape; exact tab closes discriminate by `kind`. */
-export type TabClosePayload = PaneClosePayload | TerminalTabClosePayload
-
-/** The `tab.rename` channel payload — one "name this Terminal Tab". */
-export interface TabRenamePayload {
-  readonly taskId: string
-  readonly tabId: string
-  /** The new user title. Empty clears back to the tab's default name — the
-   *  same meaning f2's rename dialog gives a blank field. */
-  readonly title: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
-}
-
-/** The `ui.prompt` channel payload — one host-dialog text-input request. */
-export interface UiPromptPayload {
-  /** Broker key the answering `ui.promptReply` names. */
-  readonly promptId: string
-  /** Dialog title (plugin-provided, shown verbatim). */
-  readonly title: string
-  readonly placeholder?: string
-  /** Pre-filled input value. */
-  readonly initial?: string
-  /** Publish time (ms epoch) — the consumer-side dedupe key. */
-  readonly at: number
 }
 
 /** The `ui-prefs` channel payload — the persisted visual prefs snapshot. */
